requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar
Keras == 2.1.6
PyYAML == 3.12
Shapely == 1.6.4.post1
astropy == 3.0.2
fact_plots == 0.3.0
h5py == 2.7.1
matplotlib == 2.2.2
numpy == 1.14.3
pandas == 0.23.0
pip install git+https://github.com/fact-project/photon_stream
pyfact == 0.19.1
scikit_learn == 0.19.1
scipy == 1.1.0
tensorflow == 1.6.0
